Sigh. Back to using unique names for files in each test.
Before this, if a test fails, all the following ones fail for
trying to create a buffer with a non-unique name.

There's another way to do this presumably using pcall to always
clean up the buffer in the test helper, but for now this is fine
again. Will clean up later.