Partially pull in teardown.
Things work differently on this branch where the check for
opening the infoview is done on each update loop, so this actually
is broken (but I broke it in an earlier commit).

I don't yet see why to do it that way (to check whether to open an
infoview) rather than just when entering new lean buffers , but
merging this in anyways since I already broke it here and will
need to rejigger update() to fix it.