Make the builtin abbreviation expansion not insert a space on <Tab>.
This is a hack, since the obvious thing (returning '') seems to not make
neovim happy -- it sits there without expanding.

Probably staring at what an autocompletion plugin does for its tab mapping
will tell us what the right way to do this is (e.g. compe) but for now this
does the job it'd appear.

Closes: #83