Make tests that don't care about specific paths not reference them.
This is basically all of them besides the one that cares about the
file contents.

I think ideally that one should be writing that specific file out
inline in the test, rather than relying on what's in the fixture
file.

But for now, this makes all the other ones just not bother knowing
where files live.