Remove lean3.init.
lean3 is a separate filetype, so this doesn't really need to exist anymore.

tree-sitter-lean is set not to require it, so just setting the filetype
on a buffer (to lean3) is enough for what this is for.

There's still a bug with properly detecting lean3 for lean3 standard library
paths (which will be fixed in a bit).