Fix infoview windows not properly setting some of their default options.
Specifically, previously infoviews did not properly :set wrap, even
though we were setting them as such.

The reason was that we setlocal some window options (like wrap, but also
winminheight and others), but we were doing so before calling :buffer
to actually put stuff in an infoview window -- so it was setting window-
local options on some random (possibly null) window.

This is likely something that neovim should warn about or document, so
will take that upstream.

For now an easy fix is just to delay when we set the filetype until we
actually are inside the window.