Remove the delay argument to Pin._update.
All the tests seem to pass without it, and some simple manual checking shows
everything still works. This may be a mistake and just isn't covered by
tests (if so someone let me know).

What I'm actually trying to do is kill this vim.wait:

https://github.com/Julian/lean.nvim/blob/2b9d10c2c1eed6f760bccd6f2fd4e46270982b6d/lua/tests/infoview/lsp/run_spec.lua#L81

and I'm on a wild goose hunt pushing it downwards with some extremely odd
behavior, which led me to this method ultimately, as pushing the vim.wait into
it (before wait_delay) was the innermost spot I can push to get the test to
pass.

Something almost certainly actually has more to do with the stateful tests in
run_spec rather than something being broken, but those tests will be rewritten.