Have the non-stateful spec files use lean.setup directly.
In the contents_spec file, this seems to somehow be related to the extra
'No info' lines being shown. Presumably there's a bug in one of the
options that the helper sets to false, but we can address that later.

For now, this saves a bit of unnecessary setup, and also these tests are
integration tests, so it seems likely more valuable that they run with
more common configurations (i.e. close to the real defaults) than that
they isolate pieces.