Rewrite the pin pausing tests.
Cuts another few seconds off, down below 22s.

This isn't per-version behavior so probably we're OK just testing it on
Lean 4. The pause portion is a bit unsettling because really we are
negatively asserting an update happens (which instead we simulate by
calling update and hoping nothing happened).

But for now it seems fine.