Remove the LSP enabled checks.
We mark this with enable = false... nil is the same as true,
so these do nothing useful. They were passing (and LSPs start)
even with require'lean'.setup{}, so this was only useful to
detect the use of the old helpers.setup{} which passed false
explicitly.

Really we need to set __index on the metatable for configs
(and use that to lookup default values).