Trigger cleanup of infoviews on their buffer, not window.
This should simplify further if/when Info disappears, but the reason
for #173 was that we were triggering cleanup of infoview state when its
window is closed, but that's not the only way an infoview can stop being
"used" to show information -- it's the infoview *buffer* which, when
visible, means the user can see what we're updating.

So, instead, this now triggers when that buffer is hidden, meaning
infoview information is no longer visible, so the infoview is
effectively closed.

Closes: #173