Reimplement assert.same.elements, fixing an autoopen bug it hid.
Unfortunately in Lua, {a, b, c} creates a table with holes
if a, b or c are nil, so doing the assertion the old way, e.g.:

    assert {a, b, c} == foo()

means it's possible for the assertion to pass because say `c` was nil
and missing from `foo()`.

This means

    assert.same.elements({a, b}, foo())

may pass accidentally, and since we essentially use this only to check
for open windows, could in theory mean a test did *not* assert a window
was open even if it looked like it did.

According to e.g.
https://sodocumentation.net/lua/topic/4475/variadic-arguments:

    Because varargs carry around their length people use them as
    sequences to avoid the issue with holes in tables. This was not
    their intended usage and one the reference implementation of Lua
    does not optimize for. Although such usage is explored in the
    examples, it is generally frowned upon.

For now, I'll take frowned upon but correct over the alternative...

Given the above (that we always use this to check windows), just
implement a `assert.windows` assertion, and use a Lua varargs trick to
know how many windows we're expecting even if some are (unintentially)
nil.

Once this is done, there was an issue with the autoopen behavior where
we indeed did not re-call options.autoopen() to check whether it returns
a new value for the current tabpage -- i.e. if:

    * we opened a window
    * autoopen() returned false
    * we opened another window on the same tabpage such that autoopen()
      now returns true

we should open a window (but previously were not).

Even after the above shpiel / cleanup, the interaction between
infoview.open and infoview.get_current_infoview():open() is really
confusing -- I want the latter to work the same as the former (and for
the former to have no reason to exist).

Doing so is easiest by reintroducing a ClosedInfoview, so maybe it's
time for that again.