Wait longer for diagnostics.
There's no harm in doing this, as it only will wait longer
on a slow machine (which mine is being right now :/).

Tests still complete in <10s.

It'd probably be tidier to refactor all the little wait
constants into some relative waits which rely on some
constant (that'd be adjustable) but that's too much work
for now.