Fix crashing when loading lean config
This fixes a crash that occurs when lean is configured outside of a
`config` function from packer. Technically---instead of communicating
changes to the filetype configuration through a shared configuration
global, the relevant configuration (default filetype) is shared through
one single global.
Furthermore, if this global is written by the user before the plugin is
loaded, then this global will not be managed by the plugin. This allows
a user to configure default filetype even if the plugin is not loaded!

Fixes #275