De-emphasize the try-this helper in the README.
It's really for Lean 3, as Lean 4 has code actiosn to do so
now.

We may remove or limit this (to Lean 3).

It does still handle by exact trimming, which it seems the code
action does not, but we can bring that up upstream if need be.

This also doesn't yet remove the mapping (which maybe someone
uses), we can also decide if/when to do that later, or if to
maybe even make the Lean 4 version of this just call the code
action automatically and keep the mapping.