And now the rest, besides files with labels which still error.
Given the latter (that these still error), we can't yet add this to
pre-commit either.

It seems there's no current way to get a reliable install for using
stylua with neovim (especially via pre-commit) -- see
JohnnyMorganz/StyLua#407 which basically says using --all-features is
broken at the minute for labels, and requires specifically compiling
only for 5.2. There doesn't seem to be a way to do this via the
pre-commit install.

The same seems true of Homebrew's stylua, which compiles with
--all-features.