Fix whatever shared state has broken the tests in 0.10.
I really don't understand what's changed here.

The only thing this does is essentially creates a new buffer for some of
the pin_spec tests (with the same contents as previous). All the rest of
the diff is whitespace noise.

But some poking around didn't really show me what was happening, and
given that I don't observe any issue when manually replicating the test,
I'm just going with this for now.