Remove support for Lean 3.
Simplifies a number of places which were doing version-specific
behavior, and blows away the Lean 3 directories.

The test suite appears to still pass, so hopefully this seems to be
correctly excised. Undoubtedly some slight further simplification is
possible in some places but I tried to be a bit conservative in this
first pass.

A `lean3` git tag is available for any future time travellers needing
Lean 3 support.