Share code between the infoview and our goal fetching commands.
This clearly removes some subtle retrying around ContentModified
errors.

It's... not clear whether that's still needed after 0.7. Let's remove it
and if we turn out to need it, it seems to belong at some higher-up
layer anyhow.

It also seems like we need to more closely package up fetching data (rpc
methods) with rendering them, which previously was inlined but here now
is (temporarily?) living in infoview.components, which we of course
never finished...

More to do.