Evade some neovim bug which causes duplicate portrait infoview windows.
I have no idea what's going on here, but this is the line which causes
it, and it does not occur for landscape infoviews.

Seems almost certainly to be a bug in winfixbuf.

To reproduce, before this change, have a portrait orientation and open a
lean file, and you should observe 3 windows opened instead of 2.

winfixbuf seems to be the culprit, and vim.schedule fixes the partial
initialization state.

Don't ask me anything further on why, that'll be for another day.