Add a first pass on showing user widgets for try this suggestions.
This effectively special-cases this widget type, but with a way of
doing so for further ones later.

Thanks to @eric-wieser for pointing out this would be possible, at least
for this case.

The (T)UI for this will likely change -- I'd like to have a telescope
preview which shows this info, but let's get it available to start.