Add an implementation of multi-selection via a popup window.
This is a sort of vim.ui.select_many interface (which doesn't exist).

Some day that may, see e.g. neovim/neovim#18161 which is an old PR
proposing it, though the PR seems essentially stale/abandoned, and
it's not clear yet what the upstream interface will be here.

In theory improvements are clearly possible to what we have, as the
nice thing about vim.ui.select is you can rig it up to a fuzzy
picker like telescope (which is less possible with this multiselect
interface. Perhaps one where that's possible is a future improvement).

For now though, this should enable our interactive hypothesis
filtering.