Trigger BufEnter when creating clean buffers in tests.
I still don't fully follow why this is so manual, the neovim API is
completely silent seemingly on what should be triggered by the various
lua API functions.

I also don't know what the right sequence is here relative to
BufWinEnter which we also trigger manually here, but which seems very
fragile towards where and when we trigger it in this function.

Doing this BufEnter would be needed for any test which doesn't move the
cursor but which wants autoupdating infoviews.

We seem actually not to have any of those at the minute
'serendipitously' -- I was about to add one for diagnostics though it
turns out we move the cursor to test that anyhow.

So really this isn't needed, but still seems worthwhile theoretically.