Trim a single trailing newline from diagnostic messages with one.
This avoids showing an infoview with trailing whitespace, which just
looks ugly.

`#check` and `#eval` are slightly frustratingly inconsistent with each
other. The former does not add a trailing newline, the latter does.

VSCode displays diagnostics the same way in these cases, regardless of
whether it had a trailing newline or not, so we do the same.

In theory we may want to strip *all* trailing newlines, but we can
decide to do that later.

(This is why the only place with failures is the autoupdate_spec file).