Fix a widget docstring and promote single-widget rendering.
We need both versions of this I guess. One which returns a `Widget` and
one which directly returns an `Element`, which we already do here (in
the `map`) but which we also will need to do immediately as part of
rendering MsgEmbeds which contain widgets which we don't support yet
but will in a second. Those contain a single user widget instance which
needs immediate rendering.