Key RPC sessions by file URI not buffer number.
The RPC sessions anyways cannot be opened simultaneously for the same
URI (RpcConnectParams take a URI, and
https://github.com/leanprover/lean4/blob/abae95e1701a081701ef132a35b3b409594e3907/src/Lean/Server/FileWorker.lean#L510
will replace a session and purge the session ID if two sessions are
created for the same URI).

This also *seems* to obviate the hack we just put in to check we have
valid bufnrs, as tests now seem to pass without those checks.