Fix showing diagnostics when the file is super broken.
The Lean server will call files permanently "processing" but the *kind*
of processing we weren't taking into account -- specifically, when a
file has gotten a `fileProgress` of kind `fatalError` it's basically not
processing. In this case we still want to show diagnostics.

For example a file which has `import TotallyBroken` will come up
"processing" in this way even though the server has given up.

This manifested in us showing an empty infoview but there were still
clearly diagnostics on the line.

In theory there's still something slightly nonsensical here, namely
that we trigger re-rendering the infoview when we get notified of
`textDocument/publishDiagnostics` -- and that request *contains*
the diagnostics, so we should possibly be passing them directly to
rerendering of the infoview. But that's failrly minor.