Another minor tweak to cleaning up our abbreviations mappings.
Loop over mappings we set rather than all the ones the user may have set
in a buffer.

It still does annoy me a bit that vim.keymap.get doesn't exist (which is
wontfixed upstream for now in neovim/neovim#24333 and which encourages
still using the vimscript API, so that's what we do).