Add some basic internal logging instrumentation.
This is useful for debugging by setting a new `log` function in a Lean
configuration.

By default, these logs go nowhere (though maybe we'll change that if we
ever add any log messages that are warnings or errors).
So as far as users are concerned this should all be invisible thus far.

To enable this then, set e.g.:

    log = function(level, message)
      if level < vim.log.levels.DEBUG then
        return
      end
      vim.print(message)
    end

in a config.

Undoubtedly there are more places it's useful to instrument -- right now
the focus is on identifying where we'll start doing some debouncing.