Implement highlighting for goal state diffing.
This adds support for two of the three kinds Lean sends:

    * showing when terms are changing
      (implementation-wise, this is from `SubexprInfo`)
    * showing when hypotheses are changing
      (implementation-wise, this is from `InteractiveHypothesisBundle`)

There's a third type we don't add support for which Lean sends (in
`InteractiveGoal`) which indicates whether whole goals are changing, and
in VSCode this is indicated via some indent-guide-like lines that appear
to the left of the goal.

We're choosing to skip that for now since it'd mean adding some sort of
similar indent-guide-y highlights. Let's wait to see if users want it
first, even though doing so should be possible by looking at however an
indent-guide plugin does it.