Ignore errors when releasing RPC sessions.
I'm not sure why / when these happen yet, as I can't reproduce quite
yet, but doing this is "safe" in the sense that on the Lean side all
that will happen is some bit of leaking memory.

Let's choose that over an error for now until we figure out what else we
should be doing (or aren't doing).