Somehow fix some errors which happen with really broken imports.
Previously Lean would occasionally complain that
`Lean.Widgets.getInteractiveDiagnostics` was an unknown RPC call.

The way to reproduce was to type `import fawoeifjowef` at the top of a
file and then start to add or remove characters from the import, which
would consistently print errors.

I really don't understand what's happening here, nor can I get this to
reproduce in an automated test, but via manual testing, this reordering
of rendering diagnostics along with *not* requesting interactive
diagnostics (and being satisfied rendering plain diagnostics when Lean
is really broken) seems to make things go away.

Given it's been a few days of knocking my head against this I'm going to
tap out here for now and put the fix in, and try to figure out what's
happening perhaps when we revisit backoff for "pin" updates.