Again simplify this pin update code.
I never understood what this did. The comment says it helps for tests,
but ... everything passes at this point with it removed.

So gone it goes.

It seems like update_pin_positions and update_pin_by_uri should be
combined, but there's a bit of for loop reworking required for that.