Remove util.subprocess_run.
vim.system exists now, which is better than vim.fn.system (it doesn't
mix stderr, which was the main issue). So it does precisely what we
need here.

We keep `subprocess_check_output` for now, as erroring on failure is
still useful to have in a helper.

This drops a plenary requirement from the module as well.