Try to silence an error coming from encoding handling.
Occasionally with changing large parts of a buffer some error is coming
out of str_utfindex.

I haven't been able to precisely reproduce yet, but the error is
annoying, so let's silence it at least for now.

This code is a vestige of our old handling of UTF32 vs. UTF16 stuff --
I'm not clear on how much of it is even really necessary anymore, but it
needs revisiting.