Do infoview diagnostic-driven updates in an autocmd not handler.
This is the more modern way to do this and also more importantly fixes
us somehow causing Satellite to think there are no diagnostics (which
itself registers a diagnostic handler this way).

I don't fully understand why, but before this change, if you had a Lean
file with some diagnostics, tracing Satellite's
diagnostic handler showed it was first getting called with the
diagnostics, and then getting called again with *empty* diagnostics,
immediately quashing the signs it would show for diagnostics (and
somehow this must have changed in the past few months).

After this, all seems fine in the world.