Remove one syn match-based highlight.
Highlight `widgetSuggestion` directly where we actually build it.

We should be reducing `syn match` / `syn region` use in favor of
explicitly assigning highlight ranges wherever possible, to avoid
needing regular expressions to know what each piece of the infoview
semantically means. (Not doing this causes issues like us
mis-highlighting a thing which looks like a goal hypothesis but is
really some random text).

We also introduce `Element.box` here as a general way to produce an
element with a title and children, though we should use this in many
more places -- let's evolve as we go, it's possible this still needs
tweaking.