Show full ranges for diagnostics in the infoview, not ranges.
Lean uses `fullRange` to indicate the full set of lines a diagnostic
applies to rather than the (LSP) range, because otherwise many lines in
the buffer would get squigglied.

When displaying the ranges in the infoview however, VSCode uses the
fullRange info to show the "real" range of the diagnostic.

Here we fix/replicate this behavior.

Thanks to @mhuisi for pointing this out!