Add RenderContext, which gives us a place to put widget rendering helpers and state.
This solves a few of the FIXMEs in the widget code and gives us a place for selection
widget helpers in the future.

Clearly there's still some interface issues to solve (particularly in tagged_text_msg_embed
which does its own call to render()) but this is a step forward.