Only stop LSP clients for the current buffer in the LSP-is-dead test.
This... fixes nothing yet unfortunately for this troublesome test, which
still seems to fail on nvim 0.11. The reason it fixes nothing is
undoubtedly because we only run one test at a time so of cousre we're
not accidentally killing other clients.

But nevertheless it seems more correct this way just in case we ever do
have any parallelism, and it's something to eliminate from diagnosing
whatever is happening here to fail the test.