Return an error from calling RPC methods on a closed buffer.
(This will fallthrough to the condition below where we then return
immediately.)

This fixes a few more tests on 0.11 -- I don't understand in what
scenario it's happening yet (and when we find out it's likely that this
check belongs a layer or two up).

But e.g. in the sorry_spec tests we seem to perhaps run so quickly that
the calls happen after the buffer is unloaded.