Bump the Lean versions for our tests to 4.19.
This involves a bunch of twiddling to insert places which now add
`Goals accomplished` to non-goal-finishing lines.

This also makes the tests a bit.. flakier :/ -- not terribly so, I can
run them ~20 times or so without typically failing. But there's a small
race condition if the test runs before the Goals accomplished markers
are populated (in which case the output will be missing that heading).

There's a few things we could do about that if it becomes an issue in
the near term:

    * make more tests *not* be full proofs. I don't know why I did this
      originally, rather than just sorrying more things, but obviously
      that would obviate places showing Goals accomplished
    * think of some way of waiting for these diagnostics to arrive, but
      I can't think of a suitable way to do so as they're push rather
      than pull
    * assert against the end of the infoview contents in some of these
      tests, but we don't have a helper that does this and it seems
      icky