Try fixing two race conditions in tests.
One happens in widget tests if the cursor moves after the buffer is
gone or closed??

And the other seems to occasionally say the infoview is empty,
presumably because diagnostics haven't arrived yet?

Let's try these two and see if they make those any less likely.

There's of course a third worse flakiness around Goals accomplished
diagnostics not arriving in time, but the "obvious" solution of trying
to wait when that happens seems not to help?!? and so needs more
thought.