Add a test for pin clicking taking you to the pin location.
This part of the codebase is so fucked, it took me 3 hours to figure out
why this test wasn't passing when run in the suite but was fine
manually, "silently" not firing the click handler even if I put
`error("ASDF")` in it.

Those 3 hours took me through trying to figure out what the heck a
path is (I still have no idea how this is updated) before figuring out
-- surely not for the first time -- that calling
`vim.api.nvim_win_set_cursor` in the test was not firing
`BufRenderer:update_cursor` and therefore the path was screwed up by the
time we fed the '<CR>' in.

All of this is beyond cursed, which I already knew. Hopefully we'll get
time to rewrite this (and brainpower to write more tests to enable
doing so without breaking it all unless I go mad writing the tests
themselves)