Show unsolved goal markers via inline, not overlay.
inline shifts the text, overlay does not.

Previously we might hide text in the buffer unintentionally when it
appears where there are unsolved goals.

The reported example (from @PatrickMassot) was:

```lean
example : ∀ n ≥ 2, True := by
  intro n : ℕ
```

where overlay would have the marker hide the colon.

(n.b. we still do this differently than VSCode which always shows the
marker at EOL. We could offer that as a configuration setting or change
the default of course.)