Don't force refreshing semantic tokens when getting file progress.
Regardless of Lean's behavior here, this seems to cause users to see
heavily flickering highlighting in buffers where many responses are
received, as Neovim first clears then re-applies semantic token
highlighting.

Let's remove this behavior here and see if instead we ever notice
semantic tokens being improperly displayed in the actual case it matters
(Where a file is long and semantic token responses for the top of the
file are sent before the bottom finishes processing).