Move unsolved goals markers to EOL.
This matches VSCode's behavior even though it's possible to have two
goals on a line where the earlier one is unsolved and the latter is not
(which makes the marker placement confusing). But we still seem to get
cases where the extmark is drawn corrupted into the buffer text, so EOL
it goes.