Again split up our fixture projects.
Specifically, don't run with a Mathlib dependency in every single test
(which we previously were doing since adding it in order to be able to
test widgets).

I still don't quite understand why but having it as a dependency causes
tests to be much slower in general, and to fail in CI on macOS often.

Let's see if this improves things (it does at least for the test run
timing).