Fix two subtle infoview contents bugs caused by Elements:concat.
We short circuit when attempting to concatenate exactly one child in an
Element.
Specifically, we just return the child.

But that's wrong if any additional customization is being done on
the element, like adding event handlers.

In practice, this means at least two things were broken:

  * we sometimes weren't properly highlighting added hypotheses
  * dismissing tooltips via <Esc> sometimes didn't work in tactic mode

We now still preserve the event handlers (and error if we have 0
children but opts has been passed).