Tweak line breaking default settings in the infoview window.
In particular, set linebreak by default to encourage breaking at
specific positions rather than simply at the window width.

Then set breakat, which controls where to break, to not contain `.`.

Of course this isn't perfect, it doesn't support unicode, and ideally
there'd be a sort of extensible soft-wrapping API we could use (and then
we could rely on Lean itself to help decide where to break), but
hopefully this is at least better than the default.