Apply winhilight to infoview windows when they're in notable states.
Effectively this changes the background color of the infoview when
attention is needed.

This applies when:

  * the LSP is dead
  * the infoview is paused

(and the idea will be to extend this to cases where imports are out of
date).

We technically support 2 different levels of attention -- a "warn"
highlight group (`leanInfoNCWarn`) used when paused -- as presumably the
user has paused the infoview and we simply want to visually indicate
this -- and then an error highlight group (`leanInfoNCError`) for when
the LSP is dead.

We only apply these groups when the infoview window is *not* the current
window (by making use of the `NormalNC` highlight group from Neovim).

Users can remap the aforementioned highlight groups to control the
precise background color. By default we use floating window highlight
groups which tend to be slight variants on the normal color in a given
theme.